Version which include checking if sba-tests_orogin is empty and in case it is, code make a copy of sba-tests to origin

+++ update_sba.py
@@ -4,7 +4,16 @@ import re
 import shutil
 source_path = 'sba-tests_origin.yaml'
 destination_path = 'sba-tests.yaml'
-shutil.copy(source_path, destination_path)
+
+
+import os
+
+def is_file_empty(file_path):
+    return os.path.exists(file_path) and os.path.getsize(file_path) == 0
+if is_file_empty(source_path):
+    shutil.copy(destination_path,source_path)
+else:
+    shutil.copy(source_path, destination_path)
 def comment_out_tags(file_path, test_name, tag_to_keep):
     with open(file_path, 'r') as file:
         lines = file.readlines()
